Simplify mediawiki authentication
This removes the seeming unnecessary overiding of request_phase
reducing the risk of diverging from the parent class version which
it was largely copied from.

It also removed the override of callback_url in favour of using
the default, which requests a callback to the correct omniauth
callback rather than requesting out-of-band callback which seems
wrong for a web authentication framework.

This has been tested - you need to make sure you specify the right
callback prefix and enable callback URLs when creating the consumer
at the mediawiki end but so long as you do that it works fine.